refactor: continue refactoring to layered architecture
