+++ backend/src/handlers/read.js
@@ -1,8 +1,8 @@
 const getVideosForUser = require("../lib/get-videos-for-user");
 const getVideo = require("../persistence/get-video");
 const getAuthToken = require("../utility/get-auth-token");
-const getTagsAvailableToUser = require("../lib/get-tags-available-to-user");
-const getEverythingInTable = require("../lib/get-everything-in-table");
+const getTagsAvailableToUser = require("../application/get-tags-available-to-user");
+const getEverythingInTable = require("../persistence/get-everything-in-table");
 const ResponseFactory = require("../lib/factories/ResponseFactory");
 const getCommentsForVideo = require("../application/get-comments-for-video");
 const isUserAnAdmin = require("../lib/is-user-an-admin");
