+++ backend/src/persistence/get-everything-in-table.js
@@ -1,5 +1,5 @@
-const DynamoFactory = require("./factories/DynamoFactory");
-const extractItemsFromDynamoResults = require("./extract-items-from-dynamo-results");
+const DynamoFactory = require("../lib/factories/DynamoFactory");
+const extractItemsFromDynamoResults = require("../lib/extract-items-from-dynamo-results");
 
 module.exports = async tableName => {
 	const items = await loadEverythingFromTable(tableName);
